Extract hashCode generation helpers to new file
